Refactored module "filter"
