+++ src/components/filter.js
@@ -1,7 +1,51 @@
 import {tasks} from "../main";
-import {generateFilters} from "../mocks/filter";
 
-const createFilterMarkup = ({name, count, isChecked}) => {
+const FILTER_NAMES = [
+  `all`,
+  `overdue`,
+  `today`,
+  `favorites`,
+  `repeating`,
+  `archive`
+];
+
+const countTodayTasks = (tasksData) => {
+  return tasksData.filter((task) => {
+    if (!task.dueDate) {
+      return false;
+    }
+    return (task.dueDate).toLocaleDateString() === (new Date()).toLocaleDateString();
+  });
+};
+
+const generateFilters = (taskData) => {
+  const quantitytByName = {
+    all: taskData,
+    overdue: taskData.filter((task) => task.dueDate < new Date()),
+    today: countTodayTasks(taskData),
+    favorites: taskData.filter((task) => task.isFavorite),
+    repeating: taskData.filter((task) => task.isRepeat),
+    archive: taskData.filter((task) => task.isArchive),
+  };
+
+  const quantityByName = FILTER_NAMES.reduce((list, name) => {
+    list[name] = quantitytByName[name].length;
+
+    return list;
+  }, {});
+
+  return FILTER_NAMES.map((name, index) => {
+    const count = quantityByName[name] || 0;
+
+    return {
+      name,
+      count,
+      isChecked: index === 0
+    };
+  });
+};
+
+export const createFilterMarkup = ({name, count, isChecked}) => {
   return (
     `<input
       type="radio"
@@ -16,7 +60,7 @@ const createFilterMarkup = ({name, count, isChecked}) => {
   );
 };
 
-const createFilterTemplate = () => {
+export const createFilterTemplate = () => {
   const filters = generateFilters(tasks);
   const filtersMarkup = filters.map(createFilterMarkup).join(`\n`);
 
@@ -24,5 +68,3 @@ const createFilterTemplate = () => {
     ${filtersMarkup}
   </section>`;
 };
-
-export {createFilterTemplate};
